[analyzer] Fix handling of annotated vars/fields in api.txt.
Changes the tool that generates `api.txt` files for the analyzer (and
related packages) so that when examining a synthetic element, it looks
for annotations on the corresponding non-synthetic element.

This fixes detection of `@deprecated` and `@experimental` on getters
and setters induced by top level variables and fields.

Addresses
https://dart-review.googlesource.com/c/sdk/+/455820/comment/f2a39140_563a86de/.

Change-Id: I6a6a6964c06905fc94b27a48a150cb2430afad99
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/456320
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>