[analysis_server] Fix Call Hierarchy target selection for tear-offs
This fixes `_findTargetNode` on function tear-offs. Previously it would take the parent node of many `SimpleIdentifiers` which in the case of a tear-off could be some random expression.

Now we only return something other than the `node` in some very specific cases where we're trying to handle constructors (so that the type name is not considered a call for named constructors, but is considered the constructor for unnamed constructors).

Fixes https://github.com/dart-lang/sdk/issues/61736

Change-Id: I815d4dc8e1f1eaadd5ee033abf9731666944e45e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/457061
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>