[pkg/vm_service] Standardize stop printing in checkRecordedStops.
If debugPrintFile and debugPrintLine are provided, provide the stop
lines to the failure expect in the same format as the recorded stops are
printed when debugPrint is true.

Add additional debug output that lists which stops were matched and
which recorded stops were skipped, and add a reason to the failure
expect that includes the expected and recorded stop indices.

TEST=ci

Change-Id: If6974a23de72a1922d64937303816c96f646199c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/457380
Commit-Queue: Tess Strickland <sstrickl@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>