[messages] Improve error reporting.
Improves the error reporting logic used by the diagnostic message code
generators so that if an exception is thrown, its message will be
prefixed by the filename, line, and column of the YAML source that led
to the error.

In the case of errors that arise from a conflict between multiple YAML
nodes (e.g., "Analyzer diagnostic name used for multiple
diagnostics"), the filename, line, and column of all conflicting YAML
nodes will be included in the error output.

Change-Id: I6a6a696442ce69e457277f88898158b62b0aa838
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/457160
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>