[messages] Add codes to `Message` objects.
Modifies the message code generation framework so that the
`analyzerCode` and `frontEndCode` values associated with diagnostic
messages are stored in the corresponding `Message` objects.

This makes it easier to check for duplicate analyzer codes because
all `Message` objects that lead to analyzer diagnostics now support an
`analyzerCode` getter. (There is a new interface these objects:
`MessageWithAnalyzerCode`.)

It also paves the way for a follow-up CL in which I plan to centralize
and clean up the logic for reporting duplicate codes. This will in
turn allow me to sanity check the analyzer `sharedName` codes.

Change-Id: I6a6a69649c751d2ff7219baf4add20742b867d09
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/457500
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>