[messages] Add sanity check for sharedName.
Adds logic to the message code generators to verify that there are no
unnecessary analyzer `sharedName` code mappings.

An unnecessary `sharedName` code mapping is an analyzer message whose
shared name is unique, but different from the analyzer code. For
example, if there were a `ParserErrorCode.FOO` message with a shared
name of `BAR`, but no other message had the shared name of `BAR`, that
would be unnecessary, because the message's analyzer code could simply
be changed to `ParserErrorCode.BAR`.

Note, however, that if there were a `ParserErrorCode.FOO` message with
a shared name of `BAR` and a `ParserErrorCode.BAR` message with no
explicitly declared `sharedName`, that would be ok, because
`ParserErrorCode.BAR` would implicitly have a shared code of `BAR`,
meaning that `BAR` would not be unique.

Change-Id: I6a6a6964da34e1950df292419231ef4db096c121
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/457541
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>