Fine. Formal parameters are not in scope of the signature of a function expression.
Specification: 9.2 Formal Parameters

The formal parameter list of a function declaration introduces a new
scope known as the function’s formal parameter scope. The formal
parameter scope △ of a non-generic function f is enclosed in the scope
where f is declared. The formal parameter scope of a generic function f
is enclosed in the type parameter scope of f. Every formal parameter
introduces a local variable into the formal parameter scope. The current
scope for the function’s signature is the scope that encloses the formal
parameter scope.

This means that in a generic function declaration, the return type and
parameter type annotations can use the formal type parameters, but the
formal parameters are not in scope in the signature.

The body of a function declaration introduces a new scope known as the
function’s body scope. The body scope of a function f is enclosed in the
scope △ introduced by the formal parameter scope of f.

Change-Id: Icc795c11f489fadd6ed25b1a5904c5dfd1601d09
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/457885
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>