[messages] Harmonize hasPublishedDocs conventions.
Changes `pkg/_fe_analyzer_shared/messages.yaml` so that every message
has a `hasPublishedDocs` entry with a value of either `true` or
`false`. Previously, only entries with a value of `true` appeared in
this file; a default value of `false` was assumed.

Also changes the code generation logic in for diagnostic messages so
that:

- It no longer assumes `false` when there is no `hasPublishedDocs`
  entry.

- The `hasPublishedDocs` field is only accessible on messages deriving
  from `MessageWithAnalyzerCode`, since these are the only messages
  for which this field makes sense.

Also changes the logic in `messages_suite.dart` so that both `false`
and `true` values are accepted for `hasPublishedDocs`.

This change brings the conventions for `hasPublishedDocs` into harmony
between `pkg/_fe_analyzer_shared/messages.yaml` and
`pkg/analyzer/messages.yaml`.

Change-Id: I6a6a6964b5e256eb9fd6bdabadbb9392f05076f0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/458071
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>