[messages] Decouple diagnosticCodeValues generation.
Refactors (a) the logic that generates analyzer message constants and
(b) the logic that generates the `diagnosticCodeValues` declaration,
so that they both are based on the pre-computed table
`DiagnosticTables.activeMessagesByPackage`.

Previously, (a) produced a list of the active messages as a side
effect, and this list was consumed by (b), so the operation of (a) and
(b) was tightly coupled.

Using `DiagnosticTables.activeMessagesByPackage` means the two
generators are loosely coupled. This paves the way for a follow-up CL
in which I will replace (a) with logic that generates the analyzer
messages as top level constants.

Change-Id: I6a6a6964f824a4443ad3a7aed119746a80623161
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/458382
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>