[messages] Simplify generation of analyzer severities.
Removes the field `GeneratedDiagnosticClassInfo.severity`. It always
matched the severity implied by the corresponding `type`, so it's
simpler and more uniform to simply base the severity on the type.

This paves the way for a follow-up CL in which I plan to move the
severity computation to a common base class.

Change-Id: I6a6a69646455fe015596d38168f7fce6f4b7eb86
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/458582
Reviewed-by: Samuel Rawlins <srawlins@google.com>