[messages] Create DiagnosticCodeImpl class.
This class is the supertype for all concrete diagnostic classes in the
analyzer, analysis server, and linter, except for `LintCode` and some
mock classes used in tests. It provides a single common implementation
of the `severity` and `type` getters, which were previously duplicated
in a large number of code-generated derived classes.

This paves the way for follow-up CLs that will simplify the
DiagnosticCode type hierarchy, removing the code-generated derived
classes entirely.

Change-Id: I6a6a696408d1a1c51883a3a652c81ed8c9fe3cb3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/458585
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>