[messages] Use named parameters in DiagnosticCode constructors.
Changes the constructors for `LintCodeWithExpectedTypes`, and the code
generated DiagnosticCode classes, so that all their constructor
parameters are named parameters.

This makes the classes in the `DiagnosticCode` hierarchy more uniform:
excluding test classes, the only remaining constructors that have
required positional parameters are now `LintCode` (which is part of
the analyzer public API and can't easily be changed), the unnamed
constructor for `LinterLintCode` (which is deprecated and will soon be
removed), and `SecurityLintCode` (which is used internally in google3
and will require more effort to change).

This change paves the way for follow-up CLs in which I will be
flattening the `DiagnosticCode` hierarchy; those follow-up CLs will
change which constructors are called by generated code, so it will be
easier to write and review them if they don't have to switch between
named and unnamed parameters.

Change-Id: I6a6a696432c8dc701fcd87a82350d185787f5d77
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/459100
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>