[private-named-parameters] Report errors if the private name has no public name.
Eventually, the analyzer and front end will need to use the
corresponding public name as the effective name of the parameter. That
means that in addition to seeing if the parameter has a public name or
not, we actually need it. Instead of calculating the public name in the
parser and then plumbing it through the formal parameter listener, I
did the error reporting separately in the front end and analyzer.

Let me know if you think that makes sense. I can also move the other
private named parameters (back) out of the parser and into the two
front ends to be consistent if you want.

Change-Id: I9ca6c003b962d9b14d8dfa46eea4b31689935f04
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/458703
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>
Auto-Submit: Bob Nystrom <rnystrom@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>