[messages] Store lint codes in Registry as DiagnosticCodes.
Changes the static type of values in `Registry._codeMap` from
`LintCode` to `DiagnosticCode`, and removes some casts that become
unnecessary as a result.

The return type of `Registry.codeForUniqueName` is similarly changed
to `DiagnosticCode?`.

This paves the way for a follow-up CL that will change the runtime
types of lint codes from `LintCode` to `DiagnosticCode`.

Change-Id: I6a6a6964262b80324ddf6b2cf445217eef521bec
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/459522
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>