[cfe] Support default values in primary constructor parameters
This adds support for default values in primary constructor parameters.

The code handling default values is refactored to ensure consistent handling of the default values, since these need to be part of the outline in some but not all cases.

The default values not included in the outline are computed during the full compilation.

Part of https://github.com/dart-lang/sdk/issues/61700

Change-Id: Ief1a516e06d1321a5b0c94b58bf33ec3837befa4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/459422
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>