[analyzer] Fix validation of optional-checks.
Fixes the logic that validates the `optional-checks` section of the
analysis options file so that it doesn't complain about the presence
of the `propagate-linter-exceptions` option (which has been supported
for a long time).

This paves the way for a follow-up CL that will turn on
`propagate-linter-exceptions` when running unit tests, so that bugs in
lints can be found more easily.

In the future, I'd like to remove the `propagate-linter-exceptions`
option entirely, and always propagate linter exceptions; I believe the
benefit of this change (helping us to find and fix linter bugs
quickly) would far exceed the risk (additional crashes seen by
customers in the event that a linter bug slips out in a release).

Change-Id: I6a6a696439166d16b95e5439f5b88efea578d6ea
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/460221
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>