[analyzer] Fix diagnostic argument mismatch in constant evaluation.
When constant evaluation of the RHS of a binary logical expression
results in an error, the diagnostic code is extracted from the
`InvalidConstant` object and placed inside an `EvaluationException`
object. Later, that `EvaluationException` object is converted back to
an `InvalidConstant`, and later reported to the user as a diagnostic.

If the diagnostic code in question requires arguments, then the
arguments must also be extracted and placed inside the
`EvaluationException` object, so that they can be propagated to the
new `InvalidContext` object and reported as a diagnostic. If this
isn't done, then the attempt to report the error as a diagnostic will
cause an exception to be thrown, since the arguments expected by the
diagnostic code won't match the actual arguments.

Note that in a follow-up CL I plan to rework this code to use the new
"literate style" of analyzer diagnostic reporting, which makes this
sort of runtime error impossible by doing the checking at compile
time. However, that's a more invasive change, so I want to make this
targeted fix first, in case we decide to cherry-pick it to the stable
branch.

Fixes https://github.com/dart-lang/sdk/issues/61761.

Change-Id: I6a6a6964e06717aea4649c367079eb3ba366dd06
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/460201
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>